backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
splunk-handler
python-json-logger
sqlalchemy
psycopg2-binary
pillow